Added timeout and more checks to TestGC
- A timeout can be set after which the test terminates automatically
- Now not only checks file size but contents as well to detect inserted
  dummy pages containing only zero-bytes

+++ src/test/TestGC.scala
@@ -1,3 +1,7 @@
+// Flashix: a verified file system for flash memory
+// (c) 2015-2016 Institute for Software & Systems Engineering <http://isse.de/flashix>
+// This code is licensed under MIT license (see LICENSE for details)
+
 package test
 
 import java.io._
@@ -9,26 +13,28 @@ import scala.collection.mutable
 
 object TestGC {
   def main(initialArgs: Array[String]) {
-    if (initialArgs.length != 3) {
+    if (initialArgs.length != 4) {
       printHelp()
       System.exit(1)
     }
 
-    val mountPoint = initialArgs(initialArgs.length - 3)
-    val pageSize = initialArgs(initialArgs.length - 2).toInt
-    val maxFilePages = initialArgs(initialArgs.length - 1).toInt
+    val mountPoint = initialArgs(initialArgs.length - 4)
+    val pageSize = initialArgs(initialArgs.length - 3).toInt
+    val maxFilePages = initialArgs(initialArgs.length - 2).toInt
+    val timeout = initialArgs(initialArgs.length - 1).toInt
     if (pageSize <= 0 || maxFilePages <= 0) {
       printHelp()
       System.exit(1)
     }
-    val t = new TestGC(mountPoint, pageSize, maxFilePages)
+    val t = new TestGC(mountPoint, pageSize, maxFilePages, timeout)
 
     t.run()
   }
 
   def printHelp() {
     println("usage:")
-    println("  testGC <mountpoint> <page_size> <maxPagesPerFile>")
+    println("  testGC <mountpoint> <page_size> <maxPagesPerFile> <timeout>")
+    println("  timeout: in seconds; <= 0 to run the test indefinitely")
   }
 }
 
@@ -39,22 +45,25 @@ object TestGC {
   * files with random sizes.
   * When started the thread will run infinitely until a file error is detected
   * or the thread is stopped manually.
-  * If possible all created files will be cleaned up at the end.
+  * If possible all created files will be cleaned up at the end (some might
+  * persist because a cleaning was not possible with a completely full device).
   *
   * Warning:
   * - may overwrite existing files
   * - may make the filesystem unusable by completely filling it
-  * @param mountPoint path, where the flashix filesystem is mounted
-  * @param pageSize the page size in the flashix filesystem
+  *
+  * @param mountPoint   path, where the flashix filesystem is mounted
+  * @param pageSize     the page size in the flashix filesystem
   * @param maxFilePages the maximum number of pages a created file should have,
   *                     larger files may be created in the process by appending
   *                     to existing ones
+  * @param timeout      if this value is > 0 then the test runs only for timeout seconds
   */
-class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runnable {
+class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int, timeout: Int = 0) extends Runnable {
   private val maxFileSize = maxFilePages * pageSize // maximum file size in bytes
   private val files = new mutable.HashMap[UUID, Int]()
   private val getRandInt = Random.IntRandomizer
-  var running = false
+  private var running = false
   private var i = 0
 
   /**
@@ -62,11 +71,12 @@ class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runna
     */
   def run(): Unit = {
     running = true
+    val startTime = System.currentTimeMillis()
 
     while (running) {
       try {
         // select a random filesystem operation
-        Math.abs(getRandInt.random() % 6) match {
+        Math.abs(getRandInt.random() % 8) match {
           case 0 => if (files.nonEmpty) removeFile(files.keySet.head)
           case 1 => if (files.nonEmpty) appendToFile(files.keySet.head)
           case 2 => if (files.nonEmpty) overwriteFile(files.keySet.head)
@@ -75,35 +85,77 @@ class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runna
         i += 1
 
         // regularly check all files for completeness
-        if (i % 256 == 0)
+        if (i % 128 == 0)
           checkAllFiles()
       } catch {
+        // a definitely invalid file has been found: terminate and cleanup
+        case e: InvalidObjectException =>
+          e.printStackTrace()
+          running = false
+
         // device may be full: remove some files and continue
         case e: IOException =>
           e.printStackTrace()
-          println("Filesystem is full. Deleting some files...")
-          for (_ <- 0 to files.keySet.size / 4)
-            removeFile(files.keySet.last)
+          partialCleanup()
 
-        // a definitely invalid file has been found: terminate and cleanup
-        // other exceptions occurred: do the same
-        case e@(_: InvalidObjectException | _: Exception) =>
+        // received an interrupt => stop running, but still do the cleanup
+        case _: InterruptedException => running = false
+
+        // other unexpected exceptions occurred: terminate after cleanup
+        case e: Exception =>
           e.printStackTrace()
-          cleanup()
           running = false
+      } finally {
+        // always check if maximum running time is reached
+        if (timeout > 0 && System.currentTimeMillis() - startTime > timeout * 1000) {
+          println("Finished test-process after timeout! No errors detected!")
+          running = false
+        }
       }
     }
 
     cleanup()
   }
 
-  private def cleanup(): Unit = {
-    println("Cleaning up all created files...")
-    for ((f, _) <- files) {
-      println(s"Cleanup: Deleting $f")
-      new File(s"$mountPoint/$f").delete()
-      files -= f
+  def stop(): Unit = {
+    running = false
+  }
+
+  /**
+    * Remove all created files from the device
+    *
+    * @param verbose only print information about progress to stdout if true
+    */
+  private def cleanup(verbose: Boolean = false): Unit = {
+    if (verbose)
+      println("Cleaning up created files...")
+
+    try {
+      for ((f, _) <- files) {
+        if (verbose)
+          println(s"Cleanup: Deleting $f")
+        new File(s"$mountPoint/$f").delete()
+        files -= f
+      }
+    } catch {
+      case _: Exception =>
+        println("Cleanup failed! Files need to be deleted manually!")
     }
+
+    if (verbose)
+      println("Cleanup complete!")
+  }
+
+  /**
+    * Delete some of the created files to free up space on a full filesystem
+    */
+  private def partialCleanup(): Unit = {
+    if (files.isEmpty)
+      return
+
+    println("Filesystem is full. Deleting some files...")
+    for (_ <- 0 to files.keySet.size / 4)
+      removeFile(files.keySet.last)
   }
 
   private def newFile(): Unit = {
@@ -111,10 +163,20 @@ class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runna
     val (name, buf) = createFileMetadata()
     println("Creating file %s with length %d" format(name, buf.length))
 
-    val out = new FileOutputStream(s"$mountPoint/$name")
-    out.write(buf)
-    out.close()
-    files(name) = buf.length
+    try {
+      val out = new FileOutputStream(s"$mountPoint/$name")
+      out.write(buf)
+      out.close()
+      files(name) = buf.length
+    } catch {
+      // file could not be written to device, most likely there is no space left
+      // => delete the file without completeness check
+      case e: IOException =>
+        e.printStackTrace()
+        files -= name
+        new File(s"$mountPoint/$name").delete()
+        partialCleanup()
+    }
   }
 
   private def createFileMetadata(): (UUID, Array[Byte]) = {
@@ -123,6 +185,11 @@ class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runna
     (f, createBuffer())
   }
 
+  /**
+    * Creates a byte-array guaranteed not to contain zero-bytes
+    *
+    * @return a byte array smaller than `maxFileSize`
+    */
   private def createBuffer(): Array[Byte] = {
     val length = if (Random.BooleanRandomizer.random()) {
       // file size may not fit exactly into page borders
@@ -132,7 +199,7 @@ class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runna
       (getRandInt.random().abs % maxFilePages) * pageSize
     }
 
-    new Array[Byte](length)
+    byteSequence().take(length).toArray
   }
 
   private def appendToFile(uuid: UUID): Unit = {
@@ -143,11 +210,19 @@ class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runna
     val newLength = files(uuid) + appendBuffer.length
     println("Appending %d Bytes to %s. New length is %d" format(appendBuffer.length, uuid, newLength))
 
-    val f = new FileOutputStream(s"$mountPoint/$uuid", true)
-    f.write(appendBuffer)
-    f.close()
+    try {
+      val f = new FileOutputStream(s"$mountPoint/$uuid", true)
+      f.write(appendBuffer)
+      f.close()
 
-    files(uuid) = newLength
+      files(uuid) = newLength
+    } catch {
+      case e: IOException =>
+        e.printStackTrace()
+        files -= uuid
+        new File(s"$mountPoint/$uuid").delete()
+        partialCleanup()
+    }
   }
 
   private def overwriteFile(uuid: UUID): Unit = {
@@ -157,63 +232,91 @@ class TestGC(mountPoint: String, pageSize: Int, maxFilePages: Int) extends Runna
     val newBuffer = createBuffer()
     println("Overwriting file %s. New length is %d" format(uuid, newBuffer.length))
 
-    val f = new FileOutputStream(s"$mountPoint/$uuid", false)
-    f.write(newBuffer)
-    f.close()
-
-    files(uuid) = newBuffer.length
+    try {
+      val f = new FileOutputStream(s"$mountPoint/$uuid", false)
+      f.write(newBuffer)
+      f.close()
+      files(uuid) = newBuffer.length
+    } catch {
+      case e: IOException =>
+        e.printStackTrace()
+        files -= uuid
+        new File(s"$mountPoint/$uuid").delete()
+        partialCleanup()
+    }
   }
 
-  private def removeFile(uuid: UUID, checkSize: Boolean = true): Unit = {
+  private def removeFile(uuid: UUID, check: Boolean = true): Unit = {
     if (files.isEmpty)
       return
 
     println(s"Deleting file $uuid")
 
+    if (check)
+      checkFile(uuid, files(uuid))
+
     val f = new File(s"$mountPoint/$uuid")
-
-    // if the device is full the size might not have been written completely,
-    // so skip the check
-    if (checkSize) {
-      val size = files(uuid)
-      val reader = new FileInputStream(f)
-      val bytesRead = reader.read(new Array(size))
-
-      if (f.length() != size || bytesRead != size) {
-        throw new InvalidObjectException(s"Actual size of file $uuid is $bytesRead, but expected it to be $size!")
-      }
-      reader.close()
-    }
-
     f.delete()
 
     files -= uuid
   }
 
   /**
-    * Check all files for completeness by reading their content and checking the length
+    * Check all files for completeness and integrity
     */
   private def checkAllFiles(): Unit = {
     println("Checking all files for completeness... ")
 
     for ((uuid, size) <- files) {
-      val f = new File(s"$mountPoint/$uuid")
-      val reader = new FileInputStream(f)
-      val bytesRead = reader.read(new Array(size))
-
-      // check that the file length reported by the OS and the readable bytes
-      // matches the written size
-      if (f.length() != size || bytesRead != size) {
-        throw new InvalidObjectException(s"Actual size of file $uuid is $bytesRead, but expected it to be $size!")
-      }
-
-      reader.close()
+      checkFile(uuid, size)
     }
 
     println("Check OK")
   }
 
-  def stop(): Unit = {
-    running = false
+  /**
+    * Check the given file for completeness and integrity
+    *
+    * Checks that the size reported by the filesystem and the readable size
+    * match the originally written size.
+    * Also checks that no "dummy pages" have been inserted to cover up missing
+    * pages. The original files have no zero-bytes, dummy pages are
+    * filled with zeroes.
+    *
+    * @param uuid filename to check
+    * @param size originally written size
+    */
+  private def checkFile(uuid: UUID, size: Int): Unit = {
+    val f = new File(s"$mountPoint/$uuid")
+    val reader = new FileInputStream(f)
+    val len = f.length().asInstanceOf[Int]
+    val fileContent = new Array[Byte](Math.max(len, size))
+    val bytesRead = reader.read(fileContent)
+
+    // check that the file length reported by the OS and the readable bytes
+    // matches the written size
+    if (f.length() != size || bytesRead != size) {
+      throw new InvalidObjectException(s"Actual size of file $uuid is $bytesRead (readable)/$len (reported by filesystem) but expected it to be $size!")
+    }
+
+    // no byte should be zero in the written files => if the read bytes contain
+    // a zero a page could not be read and we know it was given back as a
+    // placeholder
+    if (fileContent.contains(0)) {
+      throw new InvalidObjectException(s"File $uuid contains zero-bytes!")
+    }
+
+    reader.close()
   }
+
+  /**
+    * Generates a byte stream that is guaranteed not to contain zero
+    *
+    * @param start a number to count up from, default: 1
+    * @return a byte stream of non-zero bytes
+    */
+  private def byteSequence(start: Byte = 1): Stream[Byte] = start #:: byteSequence(start match {
+    case -1 => 1.asInstanceOf[Byte]
+    case _ => (start + 1).asInstanceOf[Byte]
+  })
 }
